+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Tous les changements de ce projet seront documentés dans ce fichier.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).*
 
+
+## 09-12-2021 
+## Added
+- Ajout des actions back-end pour ajouter et/ou modifier les questions
+
 ## 08-12-2021 
 ## Added
 - Modal pour ajouter et/ou modifier les questions